+++ core/src/main/java/sting/Named.java
@@ -15,10 +15,10 @@ import javax.annotation.Nonnull;
  * parameters, and {@link Injector} output methods.</p>
  *
  * <p>Sting also tolerates this annotation for framework integration on {@link InjectorFragment}
- * methods, on types or constructor parameters whose enclosing type is annotated with an
- * annotation meta-annotated by {@link StingProvider}, and on types or constructor parameters
- * whose enclosing type is annotated with an annotation meta-annotated by
- * {@link ActAsStingComponent}.</p>
+ * methods, on constructor parameters whose enclosing type is annotated with an annotation
+ * meta-annotated by {@link ActAsStingConsumer}, and on types annotated with an annotation
+ * meta-annotated by {@link ActAsStingProvider}. Types annotated with annotations meta-annotated
+ * by {@link ActAsStingComponent} receive both validation allowances.</p>
  */
 @Documented
 @Retention( RetentionPolicy.RUNTIME )