+++ processor/src/test/fixtures/bad_input/com/example/named/NamedOnCtorParamInActAsStingProviderModel.java
@@ -1,0 +1,20 @@
+package com.example.named;
+
+import sting.ActAsStingProvider;
+import sting.Named;
+
+public class NamedOnCtorParamInActAsStingProviderModel
+{
+  @ActAsStingProvider
+  @interface MyFramework
+  {
+  }
+
+  @MyFramework
+  static class MyComponentType
+  {
+    MyComponentType( @Named( "Foo" ) int i )
+    {
+    }
+  }
+}