+++ processor/src/test/fixtures/bad_input/com/example/named/NamedOnTypeInActAsStingConsumerModel.java
@@ -1,0 +1,18 @@
+package com.example.named;
+
+import sting.ActAsStingConsumer;
+import sting.Named;
+
+public class NamedOnTypeInActAsStingConsumerModel
+{
+  @ActAsStingConsumer
+  @interface MyFramework
+  {
+  }
+
+  @MyFramework
+  @Named( "Foo" )
+  static class MyComponentType
+  {
+  }
+}