+++ processor/src/test/fixtures/input/com/example/named/NamedOnTypeInActAsStingProviderModel.java
@@ -1,0 +1,18 @@
+package com.example.named;
+
+import sting.ActAsStingProvider;
+import sting.Named;
+
+public class NamedOnTypeInActAsStingProviderModel
+{
+  @ActAsStingProvider
+  @interface MyFramework
+  {
+  }
+
+  @MyFramework
+  @Named( "Foo" )
+  static class MyComponentType
+  {
+  }
+}