+++ processor/src/test/fixtures/input/com/example/provider_backed/named/NamedOnActAsStingProviderBackedTypeModel.java
@@ -1,0 +1,18 @@
+package com.example.provider_backed.named;
+
+import sting.ActAsStingProvider;
+import sting.Named;
+
+public final class NamedOnActAsStingProviderBackedTypeModel
+{
+  @ActAsStingProvider
+  @interface FrameworkComponent
+  {
+  }
+
+  @FrameworkComponent
+  @Named( "value" )
+  static class MyComponent
+  {
+  }
+}