+++ processor/src/test/fixtures/input/com/example/provider_backed/named/NamedOnCtorParamInActAsStingConsumerBackedTypeModel.java
@@ -1,0 +1,20 @@
+package com.example.provider_backed.named;
+
+import sting.ActAsStingConsumer;
+import sting.Named;
+
+public final class NamedOnCtorParamInActAsStingConsumerBackedTypeModel
+{
+  @ActAsStingConsumer
+  @interface FrameworkComponent
+  {
+  }
+
+  @FrameworkComponent
+  static class MyComponent
+  {
+    MyComponent( @Named( "value" ) final String value )
+    {
+    }
+  }
+}