+++ processor/src/test/fixtures/input/com/example/provider_backed/typed/TypedOnActAsStingComponentBackedTypeModel.java
@@ -1,0 +1,18 @@
+package com.example.provider_backed.typed;
+
+import sting.ActAsStingComponent;
+import sting.Typed;
+
+public final class TypedOnActAsStingComponentBackedTypeModel
+{
+  @ActAsStingComponent
+  @interface FrameworkComponent
+  {
+  }
+
+  @FrameworkComponent
+  @Typed( MyComponent.class )
+  static class MyComponent
+  {
+  }
+}