+++ processor/src/test/fixtures/input/com/example/provider_backed/typed/TypedOnActAsStingProviderBackedTypeModel.java
@@ -1,0 +1,18 @@
+package com.example.provider_backed.typed;
+
+import sting.ActAsStingProvider;
+import sting.Typed;
+
+public final class TypedOnActAsStingProviderBackedTypeModel
+{
+  @ActAsStingProvider
+  @interface FrameworkComponent
+  {
+  }
+
+  @FrameworkComponent
+  @Typed( MyComponent.class )
+  static class MyComponent
+  {
+  }
+}